Event Show page now  has RSVP buttons even for logged out users
Clicking the button as an unauthenticated user shows the
"Sign In" modal, which redirects to the RSVP form on successful
sign in.

Clicking through to "Sign up" instead redirects back
to the event show page on successful sign up.

Added an "if Rails.env.test?" to guard the modal's 'fade' class,
which hilariously causes capybara+poltergeist to flake. For real.
CSS3 animations mean your content will show as 'visible'
to capybara but clicking it will not achieve success. Now you know.