Use min-height instead of height for js-match-height-item resizing
This helps in situations where a js-match-height-item div gets
bigger; since the javascript doesn't know to rerun, the div
was getting cut off and looked silly. Currently this happens
due to wrapping in the actions when all three are present
(learn, volunteer, organizer console).

The longer term fix is to somehow defer the js-match-height-item
calculation until we know the element is in its final state.