Cache event role on users to reduce RSVP querying on homepage
The homepage does a bunch of queries for each event to figure
out which buttons to show the logged-in user based on whether
they're an organizer, volunteer, etc.

There is now a User#event_attendances method which gathers
all of a user RSVPs into a hash like:
5 => {
  role: Role::STUDENT,
  waitlist_position: 2
}

This structure is memoized on the user so if we need to ask
event.organizer?(user) and event.volunteer?(user) etc multiple
times in the single request it just looks up the result in
yonder hash.