Event page will only sort checked-in attendees on top _after_ event ends
Previously, during an event, the checked-in attendees would start
bubbling to the top of the attending users section, breaking the
alphabetical nature of the list.

Now, that only happens after the event is fully over, which is
also when no-shows turn gray on the list.

Also: extracted this RSVP sorting logic from the Event into an 'RsvpSorter'.