Don't let puma preload the app if there's only one worker process
It seems like some race condition causes one or both of the worker
processes to be wedged on start. This can manifest after a deploy,
or when Heroku does its daily dyno restart. The puma config now
specifies not to `preload_app!` if there's only one worker process,
because it provides no memory benefit in that situation anyway.

We may need to revisit this decision if we ever start actually
running more than one worker on each dyno.

Decrease Rack::Timeout timeout to 20 seconds