Optimize out n+1 query for waitlisted attendees on event page
Previously, the User table would get queried once for every
waitlisted student or volunteer being displayed in sidebar.

The event show page now tries to make a maximum of seven queries
for RSVPs:
* Organizers
* Confirmed Volunteer
* Checked in Volunteer
* Confirmed Student
* Checked in Student
* Waitlisted Student
* Waitlisted Volunteer

It's possible these could be better optimized by just fetching
all the RSVPs and partitioning them in Ruby, but it's hard
to know whether it would be faster overall.