Clone all existing Regions into Chapters
This is to facilitate a different meaning between 'region'
and 'chapter'.

A "region" is a collection of locations. Attendees can subscribe
to emails for events that take place in their region, and will
(currently) be mailed regardless of whether the event belongs
to "RailsBridge" or "MobileBridge"

A "chapter" represents the organizing group who caused that
event to happen. Every "event" belongs to one "chapter".
Chapters also belong to one "organization", like "RailsBridge"
or "MobileBridge"

An event now requires a Chapter to be created, but there is not
yet any UI to create one.