Convert devise/shared/_links partial into a helper method
The helper method takes a list of symbols which can be any
subset of (:sign_up, :forgot_password, :confirmation_instructions)

Directly referencing the appropriate symbols is a more explicit
than rendering the partial in a lot of diffrent contexts and
allowing it to perform logic based on the current controller.

When the logic was primarily based on the current controller, it
meant the 'forgot password?' link would not render in the sign in
modal if you were already on the Forgot Password page. Which is
arguably a "feature", but kindof weird.