Add ability to create a new location from the new events page via AJAX, eliminating the need to refresh the page.
This was done by using unobtrusive javascript, a bootstrap modal, and
code to that adds the ability of the location controller to respond to
javascript requests. The major changes are as follows:

1. The Events controller now sets an empty location for the new and
create action. This is necessary to properly render the location form.

2. The location controller now responds to both HTML and JS requests.

3. A location modal, which renders a new location form partial, has been
added to the events/_form partial if @event.new_record?.

4. A location modal partial has been created within the views/events
folder.  This modal invokes the locations/_form partial. By splitting
out the modal, we are able to only include it if @event.new_record? is
true.

5. The locations_/form partial now has its remote option set to true if
it is being rendered from events/_form.

6. Two new files, create.js.erb and create_failed.js.erb have been so
that the Location controller can respond with Javascript if the #create
action receives a javascript request.

7. A full test-suit for this feature is included.