Fix bug where re-rendering the email form on error showed incorrect recipients
The preselection was happening on RSVP ids instead of User ids,
which caused the tests to pass (because they were the same)
but would offer very bizarre results in real life.

Also: only show the 'We were unable to send your email' message
once by using flash.now instead of flash.

Also: use a SimpleDelegator for the event email presenter