Remove button styling from "Sign Up" link
I was surprised by the modal - after filling in my email and a password and clicking "Sign Up", I was taken to a longer form that didn't retain that info. I thought this would be a nice little UX stopgap, to make the link look like a link until someone wants to take a crack at the fiddly task of having it carried over.