Show childcare requests even if childcare disabled
The problem:

Oftentimes we want to provide childcare for an event but disable it
shortly before the event so that we can provide an accurate count to
the babysitter and don't need to worry about last minute signups.
Unfortunately this removes the previous childcare signup information
from the organizer console, so that once childcare is disabled we no
longer have access to the information for those who have already
requested childcare.

The solution:

This makes it so that childcare signup information is shown even if childcare
is disabled for the event. A message is shown to let the user know if
childcare is disabled.